requirements.txt
# Core dependencies
python-telegram-bot==20.7
python-dotenv==1.0.0

# Database
sqlite3

# Development dependencies (opcional)
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Testing utilities
coverage==7.3.2
pytest-cov==4.1.0

# Documentation
sphinx==7.2.6